+++ lib/rodauth/rails/feature.rb
@@ -11,7 +11,6 @@ module Rodauth
     require "rodauth/rails/feature/email"
     require "rodauth/rails/feature/instrumentation"
     require "rodauth/rails/feature/internal_request"
-    require "rodauth/rails/feature/associations"
 
     include Rodauth::Rails::Feature::Base
     include Rodauth::Rails::Feature::Callbacks
@@ -20,6 +19,5 @@ module Rodauth
     include Rodauth::Rails::Feature::Email
     include Rodauth::Rails::Feature::Instrumentation
     include Rodauth::Rails::Feature::InternalRequest
-    include Rodauth::Rails::Feature::Associations
   end
 end